+++ CotEditor/Sources/DocumentWindowController.swift
@@ -114,4 +114,20 @@ final class DocumentWindowController: NSWindowController {
 
 
 
+extension DocumentWindowController: NSUserInterfaceValidations {
+    
+    func validateUserInterfaceItem(_ item: NSValidatedUserInterfaceItem) -> Bool {
+        
+        switch item.action {
+        case #selector(showOpacitySlider)?:
+            return self.window?.styleMask.contains(.fullScreen) == false
+        default:
+            return true
+        }
+    }
+    
+}
+
+
+
 extension DocumentWindowController: NSWindowDelegate { }